Extract BeaconState interface
