+++ chain/src/main/java/org/ethereum/beacon/chain/DefaultBeaconChain.java
@@ -16,6 +16,7 @@ import org.ethereum.beacon.consensus.verifier.VerificationResult;
 import org.ethereum.beacon.core.BeaconBlock;
 import org.ethereum.beacon.core.BeaconBlocks;
 import org.ethereum.beacon.core.BeaconState;
+import org.ethereum.beacon.core.BeaconStateImpl;
 import org.ethereum.beacon.core.spec.ChainSpec;
 import org.ethereum.beacon.db.Database;
 import org.ethereum.beacon.consensus.types.Score;
@@ -53,7 +54,7 @@ public class DefaultBeaconChain implements MutableBeaconChain {
 
   private BeaconTuple initializeStorage() {
     BeaconState initialState =
-        initialTransition.apply(BeaconBlocks.createGenesis(chainSpec), BeaconState.EMPTY);
+        initialTransition.apply(BeaconBlocks.createGenesis(chainSpec), BeaconStateImpl.EMPTY);
     BeaconBlock genesis =
         BeaconBlocks.createGenesis(chainSpec).withStateRoot(initialState.getHash());
