+++ consensus/src/main/java/org/ethereum/beacon/consensus/SpecHelpers.java
@@ -42,7 +42,7 @@ public class SpecHelpers {
     UInt64 state_epoch_slot = state.getSlot().minus(state.getSlot().modulo(spec.getEpochLength()));
     assertTrue(state_epoch_slot.compareTo(slot.plus(spec.getEpochLength())) <= 0);
     assertTrue(slot.compareTo(state_epoch_slot.plus(spec.getEpochLength())) < 0);
-    return state.getShardCommitteesAtSlotsUnsafe()[safeInt(
+    return state.getShardCommitteesAtSlots()[safeInt(
         slot
             .minus(state_epoch_slot)
             .plus(spec.getEpochLength()))];
