+++ consensus/src/main/java/org/ethereum/beacon/consensus/state/InMemoryValidatorRegistryUpdater.java
@@ -5,6 +5,7 @@ import static com.google.common.base.Preconditions.checkElementIndex;
 import java.util.List;
 import java.util.Optional;
 import org.ethereum.beacon.core.BeaconState;
+import org.ethereum.beacon.core.BeaconStateImpl;
 import org.ethereum.beacon.core.operations.Deposit;
 import org.ethereum.beacon.core.operations.deposit.DepositInput;
 import org.ethereum.beacon.core.spec.ChainSpec;
@@ -107,7 +108,7 @@ public class InMemoryValidatorRegistryUpdater implements ValidatorRegistryUpdate
 
   @Override
   public BeaconState applyTo(BeaconState origin) {
-    return BeaconState.Builder.fromState(origin)
+    return BeaconStateImpl.Builder.fromState(origin)
         .withValidatorRegistry(records)
         .withValidatorBalances(balances)
         .withValidatorRegistryDeltaChainTip(deltaChainTip)
