+++ consensus/src/test/java/org/ethereum/beacon/consensus/transition/InitialStateTransitionTest.java
@@ -7,6 +7,7 @@ import java.util.List;
 import java.util.Random;
 import org.ethereum.beacon.core.BeaconBlocks;
 import org.ethereum.beacon.core.BeaconState;
+import org.ethereum.beacon.core.BeaconStateImpl;
 import org.ethereum.beacon.core.operations.Deposit;
 import org.ethereum.beacon.core.spec.ChainSpec;
 import org.ethereum.beacon.pow.DepositContract;
@@ -39,7 +40,7 @@ public class InitialStateTransitionTest {
 
     BeaconState initialState =
         initialStateTransition.apply(
-            BeaconBlocks.createGenesis(ChainSpec.DEFAULT), BeaconState.EMPTY);
+            BeaconBlocks.createGenesis(ChainSpec.DEFAULT), BeaconStateImpl.EMPTY);
 
     assertThat(initialState.getGenesisTime()).isEqualTo(genesisTime);
     assertThat(initialState.getLatestDepositRoot()).isEqualTo(receiptRoot);
